+++ serialization/php/json/src/JsonParseNode.php
@@ -99,7 +99,7 @@ class JsonParseNode implements ParseNode
             throw new InvalidArgumentException("Invalid type $type provided.");
         }
         /** @var Parsable $result */
-        $result = new ($type);
+        $result = new $type();
         if($this->onBeforeAssignFieldValues !== null) {
             $this->onBeforeAssignFieldValues($result);
         }
@@ -138,7 +138,7 @@ class JsonParseNode implements ParseNode
         if (!is_subclass_of($targetEnum, Enum::class)) {
             throw new InvalidArgumentException('Invalid enum provided.');
         }
-        return new ($targetEnum)($this->jsonNode);
+        return new $targetEnum($this->jsonNode);
     }
 
     /**