+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Fixed bugs across request adapter and serialization in PHP #1353
 - Fixed NullReferenceException in Go generator
 - Fixed incorrect mapping when the response type is `text/plain` #1356
 - Fixed a bug in Dotnet.Typescript where properties could have invalid characters #1354